requirements.txt
Faker==13.3.0
grafana-api==1.0.3
psycopg2==2.9.3
python-dotenv==0.19.2
grafanalib==0.6.2
flake8>=4.0.1,<4.1
